pyproject.toml
[project]
name = "voice-assistant-demo"
version = "0.1.0"
description = "voice assistant demo"
dependencies = [
    "fastapi>=0.115.7",
    "style-bert-vits2>=2.5.0",
    "sounddevice>=0.5.1",
    "uvicorn>=0.34.0",
    "numpy==1.26.4",
    "torch>=2.5.1",
    "torchaudio>=2.5.1",
    "torchvision>=0.20.1",
    "pyaudio>=0.2.14",
    "pyinstaller>=6.11.1",
    "speechrecognition>=3.14.1",
    "pynput>=1.7.7",
    "rich>=13.9.4",
    "openai-whisper>=20240930",
    "faster-whisper>=1.1.1",
    "httpx>=0.28.1",
    "openai>=1.60.1",
    "python-dotenv>=1.0.1",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true
